Lazy qsearch movegen and lazy killer removal
