+++ src/bm/bm_search/mod.rs
@@ -2,4 +2,3 @@ pub mod move_entry;
 pub mod move_gen;
 pub mod search;
 mod see;
-mod q_move_gen;
