+++ src/bm/bm_search/move_gen.rs
@@ -1,7 +1,7 @@
 use cozy_chess::Move;
 
 use super::move_entry::MoveEntry;
-use super::see::{compare_see, move_value};
+use super::see::{calculate_see, compare_see, move_value};
 use crate::bm::bm_util::history::History;
 use crate::bm::bm_util::history::HistoryIndices;
 use crate::bm::bm_util::position::Position;
@@ -199,6 +199,52 @@ impl OrderedMoveGen {
         if self.phase == Phase::BadCaptures {
             if let Some(index) = select_highest(&self.bad_captures, |capture| capture.score) {
                 return self.bad_captures.swap_pop(index).map(|capture| capture.mv);
+            }
+        }
+        None
+    }
+}
+
+#[derive(PartialEq, Eq)]
+enum QPhase {
+    GenCaptures,
+    GoodCaptures,
+}
+
+pub struct QSearchMoveGen {
+    phase: QPhase,
+    captures: ArrayVec<Capture, MAX_MOVES>,
+}
+
+impl QSearchMoveGen {
+    pub fn new() -> Self {
+        Self {
+            phase: QPhase::GenCaptures,
+            captures: ArrayVec::new(),
+        }
+    }
+
+    pub fn next(&mut self, pos: &Position, hist: &History) -> Option<(Move, i16)> {
+        if self.phase == QPhase::GenCaptures {
+            self.phase = QPhase::GoodCaptures;
+            let stm = pos.board().side_to_move();
+            pos.board().generate_moves(|mut piece_moves| {
+                piece_moves.to &= pos.board().colors(!stm);
+                for mv in piece_moves {
+                    let score = hist.get_capture(pos, mv) + move_value(pos.board(), mv) * 32;
+                    self.captures.push(Capture::new(mv, score));
+                }
+                false
+            });
+        }
+        if self.phase == QPhase::GoodCaptures {
+            while let Some(index) = select_highest(&self.captures, |capture| capture.score) {
+                let capture = self.captures.swap_remove(index).mv;
+                let see = calculate_see(pos.board(), capture);
+                if see < 0 {
+                    continue;
+                }
+                return Some((capture, see));
             }
         }
         None
